Refactors migration.

+++ resources/database/migrations/2013_04_12_013201_orchestra_auth_create_user_role_table.php
@@ -14,10 +14,10 @@ class OrchestraAuthCreateUserRoleTable extends Migration
     public function up()
     {
         Schema::create('user_role', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('user_id')->unsigned();
-            $table->integer('role_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('role_id')->unsigned();
 
             $table->nullableTimestamps();
